sentinel/api/requirements.txt
# Sentinel API Dependencies

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Auth & Database
supabase>=2.3.0
python-jose[cryptography]>=3.3.0

# Validation & Security
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-magic-bin>=0.4.14  # Windows compatible

# Rate Limiting
slowapi>=0.1.9

# Monitoring
structlog>=24.1.0

# HTTP Client
httpx>=0.26.0
